Update and bug fix
Updated controls so Dyad switching can be done from the trial control
window
Changed trial and rounds so that the screen changes color. When the
trial is active it is orange, when a round is running it goes red.
Various bug fixes including failure to lot button states and LED values
to file.